Fix rebase mix-ups: rename multiopython to multio
